requirements.txt
python-dotenv==1.0.0
beautifulsoup4==4.12.0
requests==2.28.2
python-docx==0.8.11
python-dotenv==1.0.0